refactor(lift-teardown): Phase B — delete Http4sLiftWebBridge, wire JSON catch-all
B1: remove Http4sLiftWebBridge.routes from Http4sApp; replace with notFoundCatchAll
    that returns {"code":404,"message":"..."} matching OBP error format.
B2: extract ensureStandardHeaders into Http4sStandardHeaders; add handleErrorWith
    500 handler in httpApp; remove bridge traffic audit route.
B3: delete Http4sLiftWebBridge.scala and the 5 bridge-specific test files
    (Http4sLiftBridgeTrafficTest, Http4sResponseConversionTest,
    Http4sCallContextBuilderTest, Http4sResponseConversionPropertyTest,
    Http4sRequestConversionPropertyTest).
    Remove wrappedRoutesV500ServicesWithBridge from Http4s500 (used only by
    the @Ignore V500ContractParityTest, updated to use WithJsonNotFound variant).

Full suite green: 2902 tests, 0 failures.
