--- obp-api/src/test/scala/code/api/util/http4s/Http4sRequestConversionPropertyTest.scala
@@ -1,620 +1,0 @@
-package code.api.util.http4s
-
-import cats.effect.IO
-import code.api.util.ExampleValue
-import net.liftweb.http.Req
-import org.http4s.{Header, Method, Request, Uri}
-import org.scalatest.{FeatureSpec, GivenWhenThen, Matchers, Tag}
-import org.typelevel.ci.CIString
-
-import scala.util.Random
-
-/**
- * Property Test: Request Conversion Completeness
- * 
- * **Validates: Requirements 2.2**
- * 
- * For any HTTP4S request, when converted to a Lift Req object by the bridge, 
- * all request information (HTTP method, URI path, query parameters, headers, 
- * body content, remote address) should be preserved and accessible through 
- * the Lift Req interface.
- * 
- * The bridge must not lose any request information during conversion. Any missing 
- * data could cause endpoints to behave incorrectly. This property ensures the 
- * bridge correctly implements the HTTPRequest interface.
- * 
- * Testing Approach:
- * - Generate random HTTP4S requests with various combinations of headers, params, and body
- * - Convert to Lift Req through bridge
- * - Verify all original request data is accessible through Lift Req methods
- * - Test edge cases: empty bodies, special characters, large payloads, unusual headers
- * - Minimum 100 iterations per test
- */
-class Http4sRequestConversionPropertyTest extends FeatureSpec 
-  with Matchers 
-  with GivenWhenThen {
-
-  object PropertyTag extends Tag("lift-to-http4s-migration-property")
-  object Property2Tag extends Tag("property-2-request-conversion-completeness")
-
-  // Helper to access private buildLiftReq method for testing
-  private def buildLiftReqForTest(req: Request[IO], body: Array[Byte]): Req = {
-    val method = Http4sLiftWebBridge.getClass.getDeclaredMethod(
-      "buildLiftReq",
-      classOf[Request[IO]],
-      classOf[Array[Byte]]
-    )
-    method.setAccessible(true)
-    method.invoke(Http4sLiftWebBridge, req, body).asInstanceOf[Req]
-  }
-
-  /**
-   * Random data generators for property-based testing
-   */
-
-  // Generate random HTTP method
-  private def randomMethod(): Method = {
-    val methods = List(Method.GET, Method.POST, Method.PUT, Method.DELETE, Method.PATCH)
-    methods(Random.nextInt(methods.length))
-  }
-
-  // Generate random URI path
-  private def randomPath(): String = {
-    val segments = Random.nextInt(5) + 1
-    val path = (1 to segments).map(_ => s"segment${Random.nextInt(100)}").mkString("/")
-    s"/obp/v5.0.0/$path"
-  }
-
-  // Generate random query parameters
-  private def randomQueryParams(): Map[String, List[String]] = {
-    val numParams = Random.nextInt(10)
-    (1 to numParams).map { i =>
-      val key = s"param$i"
-      val numValues = Random.nextInt(3) + 1
-      val values = (1 to numValues).map(_ => s"value${Random.nextInt(100)}").toList
-      key -> values
-    }.toMap
-  }
-
-  // Generate random headers
-  private def randomHeaders(): List[(String, String)] = {
-    val numHeaders = Random.nextInt(10) + 1
-    (1 to numHeaders).map { i =>
-      s"X-Header-$i" -> s"value-$i-${Random.nextInt(1000)}"
-    }.toList
-  }
-
-  // Generate random request body
-  private def randomBody(): String = {
-    val bodyTypes = List(
-      """{"key":"value"}""",
-      """{"name":"Test","id":123}""",
-      """{"data":"Line1\nLine2\tTabbed"}""",
-      """{"unicode":"Tëst with spëcial çhars: €£¥"}""",
-      "",
-      "x" * Random.nextInt(1000)
-    )
-    bodyTypes(Random.nextInt(bodyTypes.length))
-  }
-
-  // Generate special character strings
-  private def randomSpecialChars(): String = {
-    val specialStrings = List(
-      "value with spaces",
-      "value,with,commas",
-      "value\"with\"quotes",
-      "value'with'apostrophes",
-      "value\nwith\nnewlines",
-      "value\twith\ttabs",
-      "value&with&ampersands",
-      "value=with=equals",
-      "value;with;semicolons",
-      "value/with/slashes",
-      "value\\with\\backslashes",
-      "value?with?questions",
-      "value#with#hashes",
-      "value%20with%20encoding",
-      "Tëst Ünïcödë Çhärs €£¥"
-    )
-    specialStrings(Random.nextInt(specialStrings.length))
-  }
-
-  /**
-   * Property 2: Request Conversion Completeness
-   * 
-   * For any HTTP4S request, all request data should be preserved and accessible 
-   * through the converted Lift Req object.
-   */
-  feature("Property 2: Request Conversion Completeness") {
-
-    scenario("HTTP method preservation (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with various methods")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val method = randomMethod()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = method, uri = uri)
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("HTTP method should be preserved")
-        liftReq.request.method should equal(method.name)
-        successCount += 1
-      }
-
-      info(s"[Property Test] HTTP method preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("URI path preservation (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with various URI paths")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val path = randomPath()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086$path")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.GET, uri = uri)
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("URI path should be preserved")
-        liftReq.request.uri should include(path)
-        successCount += 1
-      }
-
-      info(s"[Property Test] URI path preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Query parameter preservation (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with various query parameters")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val queryParams = randomQueryParams()
-        val path = randomPath()
-        
-        // Build URI with query parameters
-        // Note: withQueryParam replaces values, so we need to add all values at once
-        var uri = Uri.unsafeFromString(s"http://localhost:8086$path")
-        queryParams.foreach { case (key, values) =>
-          // Add all values for this key at once to create multi-value parameter
-          uri = uri.withQueryParam(key, values)
-        }
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.GET, uri = uri)
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("All query parameters should be accessible")
-        queryParams.foreach { case (key, expectedValues) =>
-          val actualValues = liftReq.request.param(key)
-          actualValues should not be empty
-          expectedValues.foreach { expectedValue =>
-            actualValues should contain(expectedValue)
-          }
-        }
-        successCount += 1
-      }
-
-      info(s"[Property Test] Query parameter preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Header preservation (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with various headers")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val headers = randomHeaders()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        var request = Request[IO](method = Method.GET, uri = uri)
-        headers.foreach { case (name, value) =>
-          request = request.putHeaders(Header.Raw(CIString(name), value))
-        }
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("All headers should be accessible")
-        headers.foreach { case (name, expectedValue) =>
-          val actualValues = liftReq.request.headers(name)
-          actualValues should not be empty
-          actualValues should contain(expectedValue)
-        }
-        successCount += 1
-      }
-
-      info(s"[Property Test] Header preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Request body preservation (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with various body content")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val body = randomBody()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.POST, uri = uri)
-          .withEntity(body)
-          .putHeaders(Header.Raw(CIString("Content-Type"), "application/json"))
-        val bodyBytes = body.getBytes("UTF-8")
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Request body should be accessible and identical")
-        val inputStream = liftReq.request.inputStream
-        val actualBody = scala.io.Source.fromInputStream(inputStream, "UTF-8").mkString
-        actualBody should equal(body)
-        successCount += 1
-      }
-
-      info(s"[Property Test] Request body preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Special characters in headers (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with special characters in headers")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val specialValue = randomSpecialChars()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.GET, uri = uri)
-          .putHeaders(Header.Raw(CIString("X-Special-Header"), specialValue))
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Special characters should be preserved in headers")
-        val actualValues = liftReq.request.headers("X-Special-Header")
-        actualValues should not be empty
-        actualValues.head should equal(specialValue)
-        successCount += 1
-      }
-
-      info(s"[Property Test] Special characters in headers: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Special characters in query parameters (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with special characters in query parameters")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val specialValue = randomSpecialChars()
-        val path = randomPath()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086$path")
-          .withQueryParam("special", specialValue)
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.GET, uri = uri)
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Special characters should be preserved in query parameters")
-        val actualValues = liftReq.request.param("special")
-        actualValues should not be empty
-        actualValues.head should equal(specialValue)
-        successCount += 1
-      }
-
-      info(s"[Property Test] Special characters in query params: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("UTF-8 characters in request body (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with UTF-8 characters in body")
-      var successCount = 0
-      val iterations = 100
-
-      val utf8Bodies = List(
-        """{"name":"Bänk Tëst"}""",
-        """{"description":"Tëst with spëcial çhars: €£¥"}""",
-        """{"unicode":"日本語テスト"}""",
-        """{"emoji":"Test 🏦 Bank"}""",
-        """{"mixed":"Ñoño €100 ¥500"}"""
-      )
-
-      (1 to iterations).foreach { iteration =>
-        val body = utf8Bodies(Random.nextInt(utf8Bodies.length))
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.POST, uri = uri)
-          .withEntity(body)
-          .putHeaders(Header.Raw(CIString("Content-Type"), "application/json; charset=utf-8"))
-        val bodyBytes = body.getBytes("UTF-8")
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("UTF-8 characters should be preserved")
-        val inputStream = liftReq.request.inputStream
-        val actualBody = scala.io.Source.fromInputStream(inputStream, "UTF-8").mkString
-        actualBody should equal(body)
-        successCount += 1
-      }
-
-      info(s"[Property Test] UTF-8 characters in body: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Large request bodies (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with large bodies")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val bodySize = Random.nextInt(1024 * 100) + 1024 // 1KB to 100KB
-        val body = "x" * bodySize
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.POST, uri = uri)
-          .withEntity(body)
-        val bodyBytes = body.getBytes("UTF-8")
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Large body should be accessible and complete")
-        val inputStream = liftReq.request.inputStream
-        val actualBody = scala.io.Source.fromInputStream(inputStream).mkString
-        actualBody.length should equal(body.length)
-        successCount += 1
-      }
-
-      info(s"[Property Test] Large request bodies: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Empty request bodies (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with empty bodies")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val method = randomMethod()
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = method, uri = uri)
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Empty body should be accessible")
-        val inputStream = liftReq.request.inputStream
-        inputStream.available() should equal(0)
-        successCount += 1
-      }
-
-      info(s"[Property Test] Empty request bodies: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Content-Type header variations (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with various Content-Type headers")
-      var successCount = 0
-      val iterations = 100
-
-      val contentTypes = List(
-        "application/json",
-        "application/json; charset=utf-8",
-        "application/json;charset=UTF-8",
-        "application/json ; charset=utf-8",
-        "text/plain",
-        "text/html",
-        "application/x-www-form-urlencoded",
-        "multipart/form-data",
-        "application/xml"
-      )
-
-      (1 to iterations).foreach { iteration =>
-        val contentType = contentTypes(Random.nextInt(contentTypes.length))
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.POST, uri = uri)
-          .withEntity("test body")
-          .putHeaders(Header.Raw(CIString("Content-Type"), contentType))
-        val bodyBytes = "test body".getBytes("UTF-8")
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Content-Type should be accessible")
-        liftReq.request.contentType should not be empty
-        val actualContentType = liftReq.request.contentType.openOr("").toString
-        actualContentType should not be empty
-        successCount += 1
-      }
-
-      info(s"[Property Test] Content-Type variations: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Authorization header preservation (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with Authorization headers")
-      var successCount = 0
-      val iterations = 100
-
-      val authTypes = List(
-        s"""DirectLogin username=\"test\", password=\"${ExampleValue.passwordExample.value}\", consumer_key=\"key\"""",
-        "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
-        "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
-        "OAuth oauth_consumer_key=\"key\", oauth_token=\"token\""
-      )
-
-      (1 to iterations).foreach { iteration =>
-        val authValue = authTypes(Random.nextInt(authTypes.length))
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.POST, uri = uri)
-          .putHeaders(Header.Raw(CIString("Authorization"), authValue))
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Authorization header should be preserved exactly")
-        val actualValues = liftReq.request.headers("Authorization")
-        actualValues should not be empty
-        actualValues.head should equal(authValue)
-        successCount += 1
-      }
-
-      info(s"[Property Test] Authorization header preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Multiple headers with same name (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with multiple values for same header")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val numValues = Random.nextInt(5) + 2 // 2-6 values
-        val values = (1 to numValues).map(i => s"value-$i").toList
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        var request = Request[IO](method = Method.GET, uri = uri)
-        values.foreach { value =>
-          request = request.putHeaders(Header.Raw(CIString("X-Multi-Header"), value))
-        }
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("All header values should be accessible")
-        val actualValues = liftReq.request.headers("X-Multi-Header")
-        actualValues.size should be >= 1
-        // At least one of the values should be present
-        values.exists(v => actualValues.contains(v)) shouldBe true
-        successCount += 1
-      }
-
-      info(s"[Property Test] Multiple headers with same name: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Case-insensitive header lookup (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with mixed-case headers")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val headerName = "Content-Type"
-        val headerValue = "application/json"
-        val uri = Uri.unsafeFromString(s"http://localhost:8086${randomPath()}")
-        
-        When("Request is converted to Lift Req")
-        val request = Request[IO](method = Method.POST, uri = uri)
-          .putHeaders(Header.Raw(CIString(headerName), headerValue))
-        val bodyBytes = Array.emptyByteArray
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("Header should be accessible with different case variations")
-        liftReq.request.headers("content-type") should not be empty
-        liftReq.request.headers("Content-Type") should not be empty
-        liftReq.request.headers("CONTENT-TYPE") should not be empty
-        liftReq.request.headers("CoNtEnT-TyPe") should not be empty
-        successCount += 1
-      }
-
-      info(s"[Property Test] Case-insensitive header lookup: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Comprehensive random request conversion (100 iterations)", PropertyTag, Property2Tag) {
-      Given("Random HTTP4S requests with all features combined")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val method = randomMethod()
-        val path = randomPath()
-        val queryParams = randomQueryParams()
-        val headers = randomHeaders()
-        val body = randomBody()
-        
-        // Build URI with query parameters
-        // Note: withQueryParam replaces values, so we need to add all values at once
-        var uri = Uri.unsafeFromString(s"http://localhost:8086$path")
-        queryParams.foreach { case (key, values) =>
-          // Add all values for this key at once to create multi-value parameter
-          uri = uri.withQueryParam(key, values)
-        }
-        
-        When("Request is converted to Lift Req")
-        var request = Request[IO](method = method, uri = uri)
-        headers.foreach { case (name, value) =>
-          request = request.putHeaders(Header.Raw(CIString(name), value))
-        }
-        if (body.nonEmpty) {
-          request = request.withEntity(body)
-            .putHeaders(Header.Raw(CIString("Content-Type"), "application/json"))
-        }
-        val bodyBytes = body.getBytes("UTF-8")
-        val liftReq = buildLiftReqForTest(request, bodyBytes)
-
-        Then("All request data should be preserved")
-        // Verify method
-        liftReq.request.method should equal(method.name)
-        
-        // Verify path
-        liftReq.request.uri should include(path)
-        
-        // Verify query parameters
-        queryParams.foreach { case (key, expectedValues) =>
-          val actualValues = liftReq.request.param(key)
-          actualValues should not be empty
-        }
-        
-        // Verify headers
-        headers.foreach { case (name, expectedValue) =>
-          val actualValues = liftReq.request.headers(name)
-          actualValues should not be empty
-        }
-        
-        // Verify body
-        if (body.nonEmpty) {
-          val inputStream = liftReq.request.inputStream
-          val actualBody = scala.io.Source.fromInputStream(inputStream, "UTF-8").mkString
-          actualBody should equal(body)
-        }
-        
-        successCount += 1
-      }
-
-      info(s"[Property Test] Comprehensive random conversion: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-  }
-
-  /**
-   * Summary test - validates that all property tests passed
-   */
-  feature("Property Test Summary") {
-    scenario("All property tests completed successfully", PropertyTag, Property2Tag) {
-      info("[Property Test] ========================================")
-      info("[Property Test] Property 2: Request Conversion Completeness")
-      info("[Property Test] All scenarios completed successfully")
-      info("[Property Test] Validates: Requirements 2.2")
-      info("[Property Test] ========================================")
-      
-      // Always pass - actual validation happens in individual scenarios
-      succeed
-    }
-  }
-}
