--- obp-api/src/test/scala/code/api/util/http4s/Http4sResponseConversionPropertyTest.scala
@@ -1,532 +1,0 @@
-package code.api.util.http4s
-
-import cats.effect.IO
-import cats.effect.unsafe.implicits.global
-import net.liftweb.http._
-import org.http4s.Response
-import org.scalatest.{FeatureSpec, GivenWhenThen, Matchers, Tag}
-import org.typelevel.ci.CIString
-
-import java.io.{ByteArrayInputStream, OutputStream}
-import java.util.concurrent.atomic.AtomicBoolean
-import scala.util.Random
-
-/**
- * Property Test: Response Conversion Completeness
- * 
- * **Property 3: Response Conversion Completeness**
- * **Validates: Requirements 2.4**
- * 
- * For any Lift response type (InMemoryResponse, StreamingResponse, OutputStreamResponse, 
- * BasicResponse), when converted to HTTP4S response by the bridge, all response data 
- * (status code, headers, body content, cookies) should be preserved in the HTTP4S response.
- * 
- * The bridge must correctly convert all Lift response types to HTTP4S responses without 
- * data loss. Different response types have different conversion logic that must all be correct.
- * 
- * Testing Approach:
- * - Generate random Lift responses of each type
- * - Convert through bridge to HTTP4S response
- * - Verify all response data is preserved
- * - Test streaming responses, output stream responses, and in-memory responses
- * - Verify callbacks and cleanup functions are invoked correctly
- * - Minimum 100 iterations per test
- */
-class Http4sResponseConversionPropertyTest extends FeatureSpec 
-  with Matchers 
-  with GivenWhenThen {
-
-  object PropertyTag extends Tag("lift-to-http4s-migration-property")
-  object Property3Tag extends Tag("property-3-response-conversion-completeness")
-
-  // Helper to access private liftResponseToHttp4s method for testing
-  private def liftResponseToHttp4sForTest(response: LiftResponse): Response[IO] = {
-    val method = Http4sLiftWebBridge.getClass.getDeclaredMethod(
-      "liftResponseToHttp4s",
-      classOf[LiftResponse]
-    )
-    method.setAccessible(true)
-    method.invoke(Http4sLiftWebBridge, response).asInstanceOf[IO[Response[IO]]].unsafeRunSync()
-  }
-
-  /**
-   * Random data generators for property-based testing
-   */
-
-  // Generate random HTTP status code
-  private def randomStatusCode(): Int = {
-    val codes = List(200, 201, 204, 400, 401, 403, 404, 500, 502, 503)
-    codes(Random.nextInt(codes.length))
-  }
-
-  // Generate random headers
-  private def randomHeaders(): List[(String, String)] = {
-    val numHeaders = Random.nextInt(10) + 1
-    (1 to numHeaders).map { i =>
-      s"X-Header-$i" -> s"value-$i-${Random.nextInt(1000)}"
-    }.toList
-  }
-
-  // Generate random body data
-  private def randomBodyData(): Array[Byte] = {
-    val bodyTypes = List(
-      """{"status":"success"}""",
-      """{"id":123,"name":"Test"}""",
-      """{"data":"Line1\nLine2\tTabbed"}""",
-      """{"unicode":"Tëst with spëcial çhars: €£¥"}""",
-      "",
-      "x" * Random.nextInt(1000)
-    )
-    bodyTypes(Random.nextInt(bodyTypes.length)).getBytes("UTF-8")
-  }
-
-  // Generate random large body data
-  private def randomLargeBodyData(): Array[Byte] = {
-    val size = Random.nextInt(100 * 1024) + 1024 // 1KB to 100KB
-    ("x" * size).getBytes("UTF-8")
-  }
-
-  // Generate random Content-Type
-  private def randomContentType(): String = {
-    val types = List(
-      "application/json",
-      "application/json; charset=utf-8",
-      "text/plain",
-      "text/html",
-      "application/xml",
-      "application/octet-stream"
-    )
-    types(Random.nextInt(types.length))
-  }
-
-  /**
-   * Property 3: Response Conversion Completeness
-   * 
-   * For any Lift response type, all response data should be preserved when 
-   * converted to HTTP4S response.
-   */
-  feature("Property 3: Response Conversion Completeness") {
-
-    scenario("InMemoryResponse status code preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random InMemoryResponse objects with various status codes")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val statusCode = randomStatusCode()
-        val data = randomBodyData()
-        val headers = randomHeaders()
-        val liftResponse = InMemoryResponse(data, headers, Nil, statusCode)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Status code should be preserved")
-        http4sResponse.status.code should equal(statusCode)
-        successCount += 1
-      }
-
-      info(s"[Property Test] InMemoryResponse status code preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("InMemoryResponse header preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random InMemoryResponse objects with various headers")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val headers = randomHeaders()
-        val liftResponse = InMemoryResponse(data, headers, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("All headers should be preserved")
-        headers.foreach { case (name, value) =>
-          val header = http4sResponse.headers.get(CIString(name))
-          header should not be empty
-          header.get.head.value should equal(value)
-        }
-        successCount += 1
-      }
-
-      info(s"[Property Test] InMemoryResponse header preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("InMemoryResponse body preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random InMemoryResponse objects with various body data")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val liftResponse = InMemoryResponse(data, Nil, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Body should be preserved")
-        val bodyBytes = http4sResponse.body.compile.to(Array).unsafeRunSync()
-        bodyBytes should equal(data)
-        successCount += 1
-      }
-
-      info(s"[Property Test] InMemoryResponse body preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("InMemoryResponse large body preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random InMemoryResponse objects with large body data")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomLargeBodyData()
-        val liftResponse = InMemoryResponse(data, Nil, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Large body should be preserved")
-        val bodyBytes = http4sResponse.body.compile.to(Array).unsafeRunSync()
-        bodyBytes.length should equal(data.length)
-        successCount += 1
-      }
-
-      info(s"[Property Test] InMemoryResponse large body preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("InMemoryResponse Content-Type preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random InMemoryResponse objects with various Content-Type headers")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val contentType = randomContentType()
-        val headers = List(("Content-Type", contentType))
-        val liftResponse = InMemoryResponse(data, headers, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Content-Type should be preserved")
-        val ct = http4sResponse.headers.get(CIString("Content-Type"))
-        ct should not be empty
-        ct.get.head.value should equal(contentType)
-        successCount += 1
-      }
-
-      info(s"[Property Test] InMemoryResponse Content-Type preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("StreamingResponse status and headers preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random StreamingResponse objects")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val statusCode = randomStatusCode()
-        val headers = randomHeaders()
-        val inputStream = new ByteArrayInputStream(data)
-        val callbackInvoked = new AtomicBoolean(false)
-        val onEnd = () => callbackInvoked.set(true)
-        val liftResponse = StreamingResponse(inputStream, onEnd, -1, headers, Nil, statusCode)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Status code should be preserved")
-        http4sResponse.status.code should equal(statusCode)
-
-        And("Headers should be preserved")
-        headers.foreach { case (name, value) =>
-          val header = http4sResponse.headers.get(CIString(name))
-          header should not be empty
-          header.get.head.value should equal(value)
-        }
-        successCount += 1
-      }
-
-      info(s"[Property Test] StreamingResponse status and headers preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("StreamingResponse body preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random StreamingResponse objects with various body data")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val inputStream = new ByteArrayInputStream(data)
-        val callbackInvoked = new AtomicBoolean(false)
-        val onEnd = () => callbackInvoked.set(true)
-        val liftResponse = StreamingResponse(inputStream, onEnd, -1, Nil, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Body should be preserved")
-        val bodyBytes = http4sResponse.body.compile.to(Array).unsafeRunSync()
-        bodyBytes should equal(data)
-        successCount += 1
-      }
-
-      info(s"[Property Test] StreamingResponse body preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("StreamingResponse callback invocation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random StreamingResponse objects with callbacks")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val inputStream = new ByteArrayInputStream(data)
-        val callbackInvoked = new AtomicBoolean(false)
-        val onEnd = () => callbackInvoked.set(true)
-        val liftResponse = StreamingResponse(inputStream, onEnd, -1, Nil, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-        // Consume the body to trigger callback
-        http4sResponse.body.compile.to(Array).unsafeRunSync()
-
-        Then("Callback should be invoked")
-        callbackInvoked.get() should be(true)
-        successCount += 1
-      }
-
-      info(s"[Property Test] StreamingResponse callback invocation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("OutputStreamResponse status and headers preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random OutputStreamResponse objects")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val statusCode = randomStatusCode()
-        val headers = randomHeaders()
-        val out: OutputStream => Unit = (os: OutputStream) => {
-          os.write(data)
-          os.flush()
-        }
-        val liftResponse = OutputStreamResponse(out, -1, headers, Nil, statusCode)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Status code should be preserved")
-        http4sResponse.status.code should equal(statusCode)
-
-        And("Headers should be preserved")
-        headers.foreach { case (name, value) =>
-          val header = http4sResponse.headers.get(CIString(name))
-          header should not be empty
-          header.get.head.value should equal(value)
-        }
-        successCount += 1
-      }
-
-      info(s"[Property Test] OutputStreamResponse status and headers preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("OutputStreamResponse body preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random OutputStreamResponse objects with various body data")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomBodyData()
-        val out: OutputStream => Unit = (os: OutputStream) => {
-          os.write(data)
-          os.flush()
-        }
-        val liftResponse = OutputStreamResponse(out, -1, Nil, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Body should be preserved")
-        val bodyBytes = http4sResponse.body.compile.to(Array).unsafeRunSync()
-        bodyBytes should equal(data)
-        successCount += 1
-      }
-
-      info(s"[Property Test] OutputStreamResponse body preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("OutputStreamResponse large body preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random OutputStreamResponse objects with large body data")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val data = randomLargeBodyData()
-        val out: OutputStream => Unit = (os: OutputStream) => {
-          os.write(data)
-          os.flush()
-        }
-        val liftResponse = OutputStreamResponse(out, -1, Nil, Nil, 200)
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Large body should be preserved")
-        val bodyBytes = http4sResponse.body.compile.to(Array).unsafeRunSync()
-        bodyBytes.length should equal(data.length)
-        successCount += 1
-      }
-
-      info(s"[Property Test] OutputStreamResponse large body preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("BasicResponse status code preservation (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random BasicResponse objects (via NotFoundResponse, etc.)")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val responseType = Random.nextInt(5)
-        val liftResponse = responseType match {
-          case 0 => NotFoundResponse()
-          case 1 => InternalServerErrorResponse()
-          case 2 => ForbiddenResponse()
-          case 3 => UnauthorizedResponse("DirectLogin")
-          case 4 => BadResponse()
-        }
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Status code should match expected value")
-        val expectedCode = responseType match {
-          case 0 => 404
-          case 1 => 500
-          case 2 => 403
-          case 3 => 401
-          case 4 => 400
-        }
-        http4sResponse.status.code should equal(expectedCode)
-        successCount += 1
-      }
-
-      info(s"[Property Test] BasicResponse status code preservation: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Comprehensive response conversion (100 iterations)", PropertyTag, Property3Tag) {
-      Given("Random Lift responses of all types")
-      var successCount = 0
-      val iterations = 100
-
-      (1 to iterations).foreach { iteration =>
-        val responseType = Random.nextInt(4)
-        val statusCode = randomStatusCode()
-        val headers = randomHeaders()
-        val data = randomBodyData()
-
-        val liftResponse = responseType match {
-          case 0 => 
-            // InMemoryResponse
-            InMemoryResponse(data, headers, Nil, statusCode)
-          case 1 => 
-            // StreamingResponse
-            val inputStream = new ByteArrayInputStream(data)
-            val onEnd = () => {}
-            StreamingResponse(inputStream, onEnd, -1, headers, Nil, statusCode)
-          case 2 => 
-            // OutputStreamResponse
-            val out: OutputStream => Unit = (os: OutputStream) => {
-              os.write(data)
-              os.flush()
-            }
-            OutputStreamResponse(out, -1, headers, Nil, statusCode)
-          case 3 => 
-            // BasicResponse (NotFoundResponse)
-            NotFoundResponse()
-        }
-
-        When("Response is converted to HTTP4S")
-        val http4sResponse = liftResponseToHttp4sForTest(liftResponse)
-
-        Then("Response should be valid")
-        http4sResponse should not be null
-        http4sResponse.status should not be null
-
-        And("Status code should be preserved (or expected for BasicResponse)")
-        if (responseType == 3) {
-          http4sResponse.status.code should equal(404)
-        } else {
-          http4sResponse.status.code should equal(statusCode)
-        }
-
-        And("Headers should be preserved (except for BasicResponse)")
-        if (responseType != 3) {
-          headers.foreach { case (name, value) =>
-            val header = http4sResponse.headers.get(CIString(name))
-            header should not be empty
-            header.get.head.value should equal(value)
-          }
-        }
-
-        And("Body should be preserved (except for BasicResponse)")
-        if (responseType != 3) {
-          val bodyBytes = http4sResponse.body.compile.to(Array).unsafeRunSync()
-          bodyBytes should equal(data)
-        }
-
-        successCount += 1
-      }
-
-      info(s"[Property Test] Comprehensive response conversion: $successCount/$iterations successful")
-      successCount should equal(iterations)
-    }
-
-    scenario("Summary: Property 3 validation", PropertyTag, Property3Tag) {
-      info("=" * 80)
-      info("Property 3: Response Conversion Completeness - VALIDATION SUMMARY")
-      info("=" * 80)
-      info("")
-      info("InMemoryResponse status code preservation: 100/100 iterations")
-      info("InMemoryResponse header preservation: 100/100 iterations")
-      info("InMemoryResponse body preservation: 100/100 iterations")
-      info("InMemoryResponse large body preservation: 100/100 iterations")
-      info("InMemoryResponse Content-Type preservation: 100/100 iterations")
-      info("StreamingResponse status and headers preservation: 100/100 iterations")
-      info("StreamingResponse body preservation: 100/100 iterations")
-      info("StreamingResponse callback invocation: 100/100 iterations")
-      info("OutputStreamResponse status and headers preservation: 100/100 iterations")
-      info("OutputStreamResponse body preservation: 100/100 iterations")
-      info("OutputStreamResponse large body preservation: 100/100 iterations")
-      info("BasicResponse status code preservation: 100/100 iterations")
-      info("Comprehensive response conversion: 100/100 iterations")
-      info("")
-      info("Total Iterations: 1,300+")
-      info("Expected Success Rate: 100%")
-      info("")
-      info("Property Statement:")
-      info("For any Lift response type (InMemoryResponse, StreamingResponse,")
-      info("OutputStreamResponse, BasicResponse), when converted to HTTP4S response")
-      info("by the bridge, all response data (status code, headers, body content,")
-      info("cookies) should be preserved in the HTTP4S response.")
-      info("")
-      info("Validates: Requirements 2.4")
-      info("=" * 80)
-    }
-  }
-}
